requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2.13
dynaconf==3.1.9
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
lxml==4.9.0
protobuf==4.21.2
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pytz==2022.1
requests==2.28.1
rsa==4.8
schedule==1.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
